# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Café Run Club Rewards API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/me/dependents/{dependentId}/redemptions:
    get:
      tags: [redemptions]
      operationId: listDependentRedemptions
      summary: List redemptions for a dependent
      parameters:
        - name: dependentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Redemption'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /reward-tiers:
    get:
      tags: [rewardTiers]
      operationId: listRewardTiers
      summary: List all reward tiers
      responses:
        '200':
          description: List of reward tiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RewardTier'
    post:
      tags: [rewardTiers]
      operationId: createRewardTier
      summary: Create a reward tier (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardTierInput'
      responses:
        '201':
          description: Created reward tier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardTier'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /reward-tiers/{tierId}:
    patch:
      tags: [rewardTiers]
      operationId: updateRewardTier
      summary: Update a reward tier (admin only)
      parameters:
        - name: tierId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RewardTierUpdate'
      responses:
        '200':
          description: Updated reward tier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RewardTier'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    delete:
      tags: [rewardTiers]
      operationId: deleteRewardTier
      summary: Delete a reward tier (admin only)
      parameters:
        - name: tierId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /redemptions:
    get:
      tags: [redemptions]
      operationId: listMyRedemptions
      summary: List my redemptions
      responses:
        '200':
          description: List of redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Redemption'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [redemptions]
      operationId: requestRedemption
      summary: Request a reward redemption
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedemptionInput'
      responses:
        '201':
          description: Redemption requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redemption'
        '400':
          description: Not enough miles or already redeemed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /redemptions/all:
    get:
      tags: [redemptions]
      operationId: listAllRedemptions
      summary: List all redemptions (admin only)
      responses:
        '200':
          description: All redemptions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RedemptionWithMember'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /redemptions/{redemptionId}:
    patch:
      tags: [redemptions]
      operationId: updateRedemptionStatus
      summary: Approve or reject a redemption (admin only)
      parameters:
        - name: redemptionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RedemptionStatusUpdate'
      responses:
        '200':
          description: Updated redemption
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Redemption'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    RewardTier:
      type: object
      required: [id, name, description, milesRequired, rewardType, active]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        milesRequired:
          type: number
        rewardType:
          type: string
          enum: [coffee, smoothie, apparel, custom]
        active:
          type: boolean
    RewardTierInput:
      type: object
      required: [name, description, milesRequired, rewardType]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        milesRequired:
          type: number
          minimum: 0.1
        rewardType:
          type: string
          enum: [coffee, smoothie, apparel, custom]
        active:
          type: boolean
    RewardTierUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        milesRequired:
          type: number
          minimum: 0.1
        rewardType:
          type: string
          enum: [coffee, smoothie, apparel, custom]
        active:
          type: boolean
    Redemption:
      type: object
      required: [id, userId, rewardTierId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        rewardTierId:
          type: integer
        rewardTierName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RedemptionWithMember:
      type: object
      required: [id, userId, rewardTierId, status, createdAt, memberDisplayName]
      properties:
        id:
          type: integer
        userId:
          type: string
        rewardTierId:
          type: integer
        rewardTierName:
          type: ["string", "null"]
        memberDisplayName:
          type: string
        memberEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        adminNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RedemptionInput:
      type: object
      required: [rewardTierId]
      properties:
        rewardTierId:
          type: integer
        forUserId:
          type: string
          description: Optional userId of a dependent to request the redemption for
    RedemptionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        adminNotes:
          type: string
...